Broke out serializer package in workflow
