--- workflow/serializers_new.py
@@ -1,1199 +1,0 @@
-# -*- coding: utf-8 -*-
-import datetime
-import operator
-import dateutil.parser
-from rest_framework import serializers
-from tola.l10n_utils import l10n_date_medium, l10n_monthname
-from tola.model_utils import get_serializer
-from workflow.models import Program, SiteProfile
-from indicators.models import (
-    Indicator,
-    LevelTier,
-    Level,
-    IndicatorType,
-    Sector,
-    PeriodicTarget,
-    DisaggregationType,
-    DisaggregationLabel
-)
-from indicators.serializers_new import (
-    ProgramPageIndicatorSerializer,
-    ProgramPageIndicatorUpdateSerializer,
-    RFLevelOrderingLevelSerializer,
-    IPTTLevelSerializer,
-    IPTTTierSerializer,
-    IPTTIndicatorSerializer,
-    IPTTExcelIndicatorSerializer,
-)
-from django.db import models
-from django.utils import timezone
-from django.utils.translation import ugettext as _
-
-
-class ProgramBase:
-    """Base class for serialized program data.  Corresponds to js/models/bareProgram"""
-    results_framework = serializers.BooleanField(source='using_results_framework')
-    by_result_chain = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Program
-        fields = [
-            'pk',
-            'name',
-            'results_framework',
-            'by_result_chain',
-        ]
-
-    @classmethod
-    def _get_query_fields(cls):
-        return ['pk', 'name', '_using_results_framework', 'auto_number_indicators']
-
-    def _get_result_tier(self, program):
-        if program.using_results_framework:
-            if hasattr(program, 'prefetch_leveltiers'):
-                tiers = [tier for tier in program.prefetch_leveltiers if tier.tier_depth == 2]
-                return tiers[0].name if tiers else False
-            elif hasattr(program, 'prefetch_tiers'):
-                tiers = [tier for tier in program.prefetch_tiers if tier.tier_depth == 2]
-                return tiers[0].name if tiers else False
-            else:
-                tier = program.level_tiers.filter(tier_depth=2).first()
-                return tier.name if tier else False
-        return None
-
-
-    def get_by_result_chain(self, program):
-        """returns "by Outcome chain" or "par chaîne Résultat" for labeling the ordering filter"""
-        tier_name = self._get_result_tier(program)
-        if tier_name:
-            return _('by %(tier)s chain') % {'tier': _(tier_name)}
-        return None
-
-    def _get_program_indicators(self, program):
-        return getattr(program, 'prefetch_indicators', Indicator.rf_aware_objects.filter(program=program))
-
-    def _get_program_levels(self, program):
-        return getattr(program, 'prefetch_levels', program.levels.order_by('customsort'))
-
-    def _get_program_tiers(self, program):
-        return getattr(program, 'prefetch_tiers', program.level_tiers.order_by('tier_depth'))
-
-    def _get_levels_level_order(self, program):
-        """returns levels sorted in level order (first by depth, then by parents' order, then by customsort)"""
-        program_levels = self._get_program_levels(program)
-        levels = []
-        parents = [None]
-        child_levels = [l for l in program_levels if l.parent_id is None]
-        while child_levels:
-            levels += child_levels
-            parents = child_levels
-            child_levels = [l for parent in parents for l in program_levels if l.parent_id == parent.pk]
-        return levels
-
-    def _get_level_children(self, parent, level_set):
-        """recursively maps levels children and their children to produce a chain-sorted level list"""
-        levels = []
-        child_levels = [l for l in level_set if l.parent_id == (parent.pk if parent else None)]
-        for child_level in child_levels:
-            levels.append(child_level)
-            levels += self._get_level_children(child_level, level_set)
-        return levels
-
-    def _get_levels_chain_order(self, program):
-        """returns levels sorted in chain order (each parent and their children and their children)"""
-        program_levels = self._get_program_levels(program)
-        levels = self._get_level_children(None, program_levels)
-        return levels
-
-    def _get_unassigned_indicators(self, indicator_set):
-        return sorted(
-            sorted(
-                [i for i in indicator_set if i.level_id is None or i.results_framework is False],
-                key=lambda i: i.sort_number
-                ),
-            key=lambda i: (i.old_level_pk is None, i.old_level_pk)
-        )
-
-ProgramBaseSerializer = get_serializer(ProgramBase)
-
-
-class ProgramReportingPeriodMixin:
-    """Extends the base program serializer to include date information, corresponds to js/models/withReportingPeriod"""
-    reporting_period_start_iso = serializers.DateField(source='reporting_period_start')
-    reporting_period_end_iso = serializers.DateField(source='reporting_period_end')
-    percent_complete = serializers.SerializerMethodField()
-    has_started = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'start_date',
-            'end_date',
-            'reporting_period_start_iso',
-            'reporting_period_end_iso',
-            'percent_complete',
-            'has_started'
-        ]
-
-    @classmethod
-    def _get_query_fields(cls):
-        return super(ProgramReportingPeriodMixin, cls)._get_query_fields() + [
-            'start_date',
-            'end_date',
-            'reporting_period_start',
-            'reporting_period_end'
-        ]
-
-
-    def get_percent_complete(self, program):
-        if program.reporting_period_end is None or program.reporting_period_start is None:
-            return -1 # otherwise the UI might show "None% complete"
-        if program.reporting_period_start > datetime.datetime.utcnow().date():
-            return 0
-        total_days = (program.reporting_period_end - program.reporting_period_start).days
-        complete = (datetime.datetime.utcnow().date() - program.reporting_period_start).days
-        return int(round(float(complete)*100/total_days)) if complete < total_days else 100
-
-    def get_has_started(self, program):
-        return program.reporting_period_start and program.reporting_period_start <= datetime.date.today()
-
-
-ProgramReportingPeriodSerializer = get_serializer(ProgramReportingPeriodMixin, ProgramBase)
-
-
-class Period:
-
-    @classmethod
-    def lop_period(cls):
-        return cls(1, {
-            'name': _('Life of Program')
-        }, True)
-
-    def __init__(self, frequency, period, tva=False):
-        self.period = period
-        self.frequency = int(frequency)
-        self.tva = tva
-
-    @property
-    def period_count(self):
-        return self.period['customsort']
-
-    @property
-    def columns(self):
-        actual_column = {
-            'header': _('Actual'),
-            'attribute': self.actual_attribute,
-            'disaggregation_attribute': self.actual_disaggregation_attribute
-        }
-        if not self.tva:
-            return [actual_column]
-        else:
-            return [
-                {
-                    'header': _('Target'),
-                    'attribute': self.target_attribute,
-                    'disaggregation_attribute': None
-                 },
-                actual_column,
-                {
-                    'header': _('% Met'),
-                    'attribute': self.percent_met_attribute,
-                    'disaggregation_attribute': None
-                }
-            ]
-    @property
-    def target_attribute(self):
-        if self.frequency == 1:
-            return 'lop_target_real'
-        else:
-            return 'frequency_{f}_period_{p}_target'.format(f=self.frequency, p=self.period_count)
-
-    @property
-    def actual_attribute(self):
-        if self.frequency == 1:
-            return 'lop_actual'
-        else:
-            return 'frequency_{f}_period_{p}'.format(f=self.frequency, p=self.period_count)
-
-    @property
-    def percent_met_attribute(self):
-        if self.frequency == 1:
-            return 'lop_met_target_decimal'
-
-    @property
-    def actual_disaggregation_attribute(self):
-        if self.frequency == 1:
-            return 'disaggregation_{}_lop_actual'
-        else:
-            return 'disaggregation_{}_' + 'frequency_{f}_period_{p}'.format(f=self.frequency, p=self.period_count)
-
-    @property
-    def start_display(self):
-        return l10n_date_medium(self.period['start'], decode=True)
-
-    @property
-    def end_display(self):
-        return l10n_date_medium(self.period['end'], decode=True)
-
-    @property
-    def header(self):
-        """LOP/MidEnd/Monthly have no header, just the label, time-aware has the period name"""
-        if self.frequency in [1, 2, 7]:
-            return None
-        return self.period['name']
-
-    @property
-    def subheader(self):
-        """ Name for Lop/MidEnd/monthly is the subheader on the excel report, label for other time-aware"""
-        if self.frequency in [1, 2, 7]:
-            return self.period['name']
-        return self.period['label']
-
-
-class ProgramLevelOrderingMixin:
-    """Extends program serializer to include a list of indicator pks in level order and chain order"""
-    indicator_pks_level_order = serializers.SerializerMethodField()
-    indicator_pks_chain_order = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'indicator_pks_level_order',
-            'indicator_pks_chain_order'
-        ]
-
-    def _get_indicators_for_ordering(self, program):
-        return Indicator.rf_aware_objects.filter(program_id=program.pk)
-
-    def get_indicator_pks_level_order(self, program):
-        """returns a list of pks, sorted by level in level order then indicator level_order (if RF) otherwise manual"""
-        indicator_pks = []
-        indicators = self._get_indicators_for_ordering(program)
-        levels = self._get_levels_level_order(program) if program.results_framework else []
-        for level in levels:
-            indicator_pks += [i.pk for i in sorted(
-                indicators, key=operator.attrgetter('level_order')
-                ) if i.level_id == level.pk]
-        indicator_pks += [i.pk for i in self._get_unassigned_indicators(indicators)]
-        return indicator_pks
-
-    def get_indicator_pks_chain_order(self, program):
-        """returns a list of pks, sorted by level in chain order then indicator level_order (if RF) otherwise manual"""
-        indicator_pks = []
-        indicators = self._get_indicators_for_ordering(program)
-        levels = self._get_levels_chain_order(program) if program.results_framework else []
-        for level in levels:
-            indicator_pks += [i.pk for i in sorted(
-                indicators, key=operator.attrgetter('level_order')
-                ) if i.level_id == level.pk]
-        indicator_pks += [i.pk for i in self._get_unassigned_indicators(indicators)]
-        return indicator_pks
-
-
-ProgramLevelOrderingProgramSerializer = get_serializer(
-    ProgramLevelOrderingMixin,
-    ProgramBase
-)
-
-
-class RFLevelOrderingMixin:
-    """Extends program serializer to include a list of level pks in level order and chain order
-
-        also adds serialized level data to program for IPTT Report
-    """
-    level_pks_level_order = serializers.SerializerMethodField()
-    level_pks_chain_order = serializers.SerializerMethodField()
-    indicator_pks_for_level = serializers.SerializerMethodField()
-    unassigned_indicator_pks = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'level_pks_level_order',
-            'level_pks_chain_order',
-            'indicator_pks_for_level',
-            'unassigned_indicator_pks'
-        ]
-
-    def _get_indicators_for_ordering(self, program):
-        return getattr(program, 'prefetch_indicators',
-                       Indicator.rf_aware_objects.filter(program_id=program.pk))
-
-    def get_level_pks_level_order(self, program):
-        return [l.pk for l in self._get_levels_level_order(program)]
-
-    def get_level_pks_chain_order(self, program):
-        return [l.pk for l in self._get_levels_chain_order(program)]
-
-    def get_indicator_pks_for_level(self, program):
-        context = {'indicators': self._get_indicators_for_ordering(program)}
-        return RFLevelOrderingLevelSerializer(
-            self._get_program_levels(program), context=context, many=True
-            ).data
-
-    def get_unassigned_indicator_pks(self, program):
-        return [i.pk for i in self._get_unassigned_indicators(self._get_indicators_for_ordering(program))]
-
-
-RFLevelOrderingProgramSerializer = get_serializer(
-    RFLevelOrderingMixin,
-    ProgramBase
-)
-
-
-class ProgramLevelUpdateMixin:
-    class Meta:
-        fields = []
-
-    @classmethod
-    def update_ordering(cls, pk):
-        program = Program.rf_aware_objects.select_related(None).prefetch_related(None).only(
-            *cls._get_query_fields()
-        ).prefetch_related(
-            models.Prefetch(
-                'indicator_set',
-                queryset=Indicator.program_page_objects.select_related(None).prefetch_related(None).only(
-                    'pk', 'name', 'deleted', 'program_id', 'level_id', 'level_order', 'number', 'old_level'
-                ),
-                to_attr='prefetch_indicators'
-            ),
-            models.Prefetch(
-                'level_tiers',
-                queryset=LevelTier.objects.select_related(None).only(
-                    'pk', 'name', 'program_id', 'tier_depth'
-                ),
-                to_attr='prefetch_leveltiers'
-            ),
-            models.Prefetch(
-                'levels',
-                queryset=Level.objects.select_related(None).only(
-                    'pk', 'name', 'parent_id', 'customsort', 'program_id'
-                ),
-                to_attr='prefetch_levels'
-            )
-        ).get(pk=pk)
-        return cls(program)
-
-
-ProgramLevelUpdateSerializer = get_serializer(
-    ProgramLevelUpdateMixin,
-    RFLevelOrderingMixin,
-    ProgramBase
-)
-
-
-class ProgramPageMixin:
-    needs_additional_target_periods = serializers.BooleanField()
-    site_count = serializers.SerializerMethodField()
-    has_levels = serializers.SerializerMethodField()
-    gait_url = serializers.CharField()
-    target_period_info = serializers.SerializerMethodField()
-    indicators = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'needs_additional_target_periods',
-            'site_count',
-            'has_levels',
-            'gait_url',
-            'target_period_info',
-            'indicators'
-        ]
-
-    @classmethod
-    def _get_query_fields(cls):
-        return super(ProgramPageMixin, cls)._get_query_fields() + ['gaitid']
-
-    @classmethod
-    def get_for_pk(cls, pk):
-        program = Program.program_page_objects.select_related(None).prefetch_related(None).only(
-            *cls._get_query_fields()
-        ).annotate(
-            num_sites=models.Count('indicator__result__site', distinct=True)
-        ).prefetch_related(
-            models.Prefetch(
-                'indicator_set',
-                queryset=Indicator.program_page_objects.select_related(None).prefetch_related(None).only(
-                    'pk', 'name', 'deleted', 'program_id', 'means_of_verification', 'level_id', 'level_order',
-                    'number', 'target_frequency', 'unit_of_measure', 'unit_of_measure_type', 'baseline', 'baseline_na',
-                    'direction_of_change', 'is_cumulative', 'key_performance_indicator', 'old_level',
-                    'create_date'
-                ),
-                to_attr='program_page_indicators'
-            ),
-            models.Prefetch(
-                'level_tiers',
-                queryset=LevelTier.objects.select_related(None).only(
-                    'pk', 'name', 'program_id', 'tier_depth'
-                ),
-                to_attr='prefetch_leveltiers'
-            ),
-            models.Prefetch(
-                'levels',
-                queryset=Level.objects.select_related(None).only(
-                    'pk', 'name', 'parent_id', 'customsort', 'program_id'
-                ),
-                to_attr='prefetch_levels'
-            )
-        ).get(pk=pk)
-        return cls(program)
-
-    def _get_indicators_for_ordering(self, program):
-        return getattr(
-            program, 'program_page_indicators',
-            super(ProgramPageMixin, self)._get_indicators_for_ordering(program)
-            )
-
-    def _get_program_levels(self, program):
-        return sorted(
-            getattr(
-                program, 'prefetch_levels',
-                super(ProgramPageMixin, self)._get_program_levels(program)
-                ), key=operator.attrgetter('customsort')
-            )
-
-    def get_site_count(self, program):
-        if hasattr(program, 'num_sites'):
-            return program.num_sites
-        return len(program.get_sites())
-
-    def get_has_levels(self, program):
-        if hasattr(program, 'prefetch_levels'):
-            return len(program.prefetch_levels)
-        return program.levels.count()
-
-    def get_target_period_info(self, program):
-        indicators = self._get_indicators_for_ordering(program)
-        irregulars = {
-            frequency: len([i.pk for i in indicators
-                            if i.target_frequency == frequency]) > 0
-            for frequency in Indicator.IRREGULAR_TARGET_FREQUENCIES
-        }
-        regulars = {
-            frequency: [i.most_recent_completed_target_end_date for i in indicators
-                        if (i.target_frequency == frequency and
-                            i.most_recent_completed_target_end_date is not None)]
-            for frequency in Indicator.REGULAR_TARGET_FREQUENCIES
-        }
-        return {
-            'lop': irregulars[Indicator.LOP],
-            'midend': irregulars[Indicator.MID_END],
-            'event': irregulars[Indicator.EVENT],
-            'time_targets': any(regulars.values()),
-            'annual': max(regulars[Indicator.ANNUAL]).isoformat() if regulars[Indicator.ANNUAL] else None,
-            'semi_annual': max(
-                regulars[Indicator.SEMI_ANNUAL]
-                ).isoformat() if regulars[Indicator.SEMI_ANNUAL] else None,
-            'tri_annual': max(regulars[Indicator.TRI_ANNUAL]).isoformat() if regulars[Indicator.TRI_ANNUAL] else None,
-            'quarterly': max(regulars[Indicator.QUARTERLY]).isoformat() if regulars[Indicator.QUARTERLY] else None,
-            'monthly': max(regulars[Indicator.MONTHLY]).isoformat() if regulars[Indicator.MONTHLY] else None,
-        }
-
-    def get_indicators(self, program):
-        return {
-            serialized_indicator['pk']: serialized_indicator
-            for serialized_indicator in ProgramPageIndicatorSerializer(
-                getattr(
-                    program, 'program_page_indicators',
-                    Indicator.program_page_objects.filter(program_id=program.pk)
-                    ),
-                many=True).data
-            }
-
-
-ProgramPageProgramSerializer = get_serializer(
-    ProgramPageMixin,
-    ProgramLevelOrderingMixin,
-    ProgramReportingPeriodMixin,
-    ProgramBase
-)
-
-class ProgramPageUpdateMixin:
-    indicator = serializers.SerializerMethodField()
-    indicators = serializers.SerializerMethodField()
-
-    class Meta:
-        override_fields = True
-        fields = [
-            'pk',
-            'indicator_pks_level_order',
-            'indicator_pks_chain_order',
-            'indicator',
-            'indicators'
-        ]
-
-    @classmethod
-    def _get_query_fields(cls):
-        return ['pk', '_using_results_framework', 'auto_number_indicators',
-                'reporting_period_start', 'reporting_period_end']
-
-    @classmethod
-    def update_indicator_pk(cls, pk, indicator_pk):
-        program = Program.program_page_objects.select_related(None).prefetch_related(None).only(
-            *cls._get_query_fields()
-        ).annotate(
-            ordering_only=models.Value(False, output_field=models.BooleanField()),
-            indicator_update_id=models.Value(indicator_pk, output_field=models.IntegerField())
-        ).prefetch_related(
-            models.Prefetch(
-                'indicator_set',
-                queryset=Indicator.program_page_objects.filter(
-                    pk=indicator_pk
-                ).select_related(None).prefetch_related(None).only(
-                    'pk', 'name', 'deleted', 'program_id', 'means_of_verification', 'level_id', 'level_order',
-                    'number', 'target_frequency', 'unit_of_measure', 'unit_of_measure_type', 'baseline', 'baseline_na',
-                    'direction_of_change', 'is_cumulative', 'key_performance_indicator', 'old_level',
-                    'create_date'
-                ),
-                to_attr='program_page_indicators'
-            ),
-            models.Prefetch(
-                'indicator_set',
-                queryset=Indicator.rf_aware_objects.select_related(None).prefetch_related(None).only(
-                    'pk', 'program_id', 'level_id', 'level_order', 'number', 'old_level',
-                ),
-                to_attr='ordering_indicators'
-            ),
-            models.Prefetch(
-                'level_tiers',
-                queryset=LevelTier.objects.select_related(None).only(
-                    'pk', 'name', 'program_id', 'tier_depth'
-                ),
-                to_attr='prefetch_leveltiers'
-            ),
-            models.Prefetch(
-                'levels',
-                queryset=Level.objects.select_related(None).only(
-                    'pk', 'name', 'parent_id', 'customsort', 'program_id'
-                ),
-                to_attr='prefetch_levels'
-            )
-        ).get(pk=pk)
-        return cls(program)
-
-    @classmethod
-    def update_ordering(cls, pk):
-        program = Program.program_page_objects.select_related(None).prefetch_related(None).only(
-            *cls._get_query_fields()
-        ).annotate(
-            ordering_only=models.Value(True, output_field=models.BooleanField())
-        ).prefetch_related(
-            models.Prefetch(
-                'indicator_set',
-                queryset=Indicator.rf_aware_objects.select_related(None).prefetch_related(None).only(
-                    'pk', 'program_id', 'level_id', 'level_order', 'number', 'old_level',
-                ),
-                to_attr='ordering_indicators'
-            ),
-            models.Prefetch(
-                'level_tiers',
-                queryset=LevelTier.objects.select_related(None).only(
-                    'pk', 'name', 'program_id', 'tier_depth'
-                ),
-                to_attr='prefetch_leveltiers'
-            ),
-            models.Prefetch(
-                'levels',
-                queryset=Level.objects.select_related(None).only(
-                    'pk', 'name', 'parent_id', 'customsort', 'program_id'
-                ),
-                to_attr='prefetch_levels'
-            )
-        ).get(pk=pk)
-        return cls(program)
-
-    def _get_program_levels(self, program):
-        return sorted(
-            getattr(
-                program, 'prefetch_levels',
-                super(ProgramPageUpdateMixin, self)._get_program_levels(program)
-                ), key=operator.attrgetter('customsort')
-            )
-
-    def _get_indicators_for_ordering(self, program):
-        return getattr(
-            program, 'ordering_indicators',
-            super(ProgramPageUpdateMixin, self)._get_indicators_for_ordering(program)
-            )
-
-    def get_indicator(self, program):
-        if getattr(program, 'ordering_only'):
-            return []
-        return ProgramPageIndicatorSerializer(
-            getattr(
-                program, 'program_page_indicators',
-                Indicator.program_page_objects.filter(pk=program.indicator_update_id)
-            )[0]).data
-
-    def get_indicators(self, program):
-        return {
-            serialized_indicator['pk']: serialized_indicator
-            for serialized_indicator in ProgramPageIndicatorUpdateSerializer(
-                getattr(
-                    program, 'ordering_indicators',
-                    Indicator.program_page_objects.filter(program_id=program.pk)
-                    ),
-                many=True).data
-            }
-
-ProgramPageUpdateSerializer = get_serializer(
-    ProgramPageUpdateMixin,
-    ProgramLevelOrderingMixin,
-    ProgramReportingPeriodMixin,
-    ProgramBase
-)
-
-class QSPeriodDateRangeSerializer(serializers.Serializer):
-    past = serializers.SerializerMethodField()
-
-    def get_past(self, period):
-        return period['start'] < self.context.get('now', timezone.now().date())
-
-
-class PeriodDateRangeSerializer(serializers.Serializer):
-    start = serializers.DateField()
-    end = serializers.DateField()
-    name = serializers.SerializerMethodField()
-    label = serializers.CharField()
-    start_label = serializers.SerializerMethodField()
-    end_label = serializers.SerializerMethodField()
-    past = serializers.SerializerMethodField()
-    year = serializers.SerializerMethodField()
-
-    def get_start_label(self, period):
-        return l10n_date_medium(period['start'], decode=True)
-
-    def get_end_label(self, period):
-        return l10n_date_medium(period['end'], decode=True)
-
-    def get_past(self, period):
-        return period['start'] < timezone.now().date()
-
-    def get_name(self, period):
-        if self.context.get('frequency') == 7:
-            return l10n_monthname(period['start'], decode=True)
-        return _(period['name'])
-
-    def get_year(self, period):
-        return period['start'].year
-
-
-class ProgramPeriodsMixin:
-    frequencies = serializers.SerializerMethodField()
-    period_date_ranges = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'frequencies',
-            'period_date_ranges'
-        ]
-
-    def get_frequencies(self, program):
-        return sorted(
-            set(i.target_frequency for i in self._get_program_indicators(program) if i.target_frequency is not None)
-        )
-
-
-    def get_period_date_ranges(self, program):
-        return {
-            frequency: PeriodDateRangeSerializer(
-                list(program.get_periods_for_frequency(frequency)),
-                many=True, context={'frequency': frequency, 'now': self.context.get('now', timezone.now().date())}
-            ).data
-            for frequency, _ in Indicator.TARGET_FREQUENCIES if frequency != Indicator.EVENT
-        }
-
-
-class IPTTQSMixin:
-    class Meta:
-        override_fields = True
-        fields = [
-            'pk',
-            'name',
-            'start_date',
-            'end_date',
-            'reporting_period_start_iso',
-            'reporting_period_end_iso',
-            'has_started',
-            'frequencies',
-            'period_date_ranges'
-        ]
-
-    @classmethod
-    def _get_query_fields(cls):
-        return ['pk', 'name', 'reporting_period_start', 'reporting_period_end', 'start_date', 'end_date']
-
-    @classmethod
-    def load_for_user(cls, user):
-        return cls.load_for_pks(user.tola_user.available_programs.annotate(
-            targets_exist=models.Exists(
-                PeriodicTarget.objects.filter(
-                    indicator__program=models.OuterRef('pk')
-                )
-            )
-        ).filter(
-            funding_status="Funded",
-            targets_exist=True,
-            reporting_period_start__isnull=False, reporting_period_end__isnull=False
-        ).values_list('id', flat=True))
-
-    @classmethod
-    def load_for_pks(cls, pks):
-        now = timezone.now().date()
-        programs = Program.rf_aware_objects.only(
-            *cls._get_query_fields()
-        ).prefetch_related(
-            models.Prefetch(
-                'indicator_set',
-                queryset=Indicator.objects.order_by().select_related(None).prefetch_related(None).only(
-                    'pk', 'program_id', 'target_frequency'
-                ).filter(program_id__in=pks),
-                to_attr='prefetch_indicators'
-            )
-        ).filter(pk__in=pks)
-        return cls(programs, many=True, context={'now': now})
-
-    def get_period_date_ranges(self, program):
-        return {
-            frequency: QSPeriodDateRangeSerializer(
-                list(program.get_short_form_periods_for_frequency(frequency)), many=True,
-                context={'frequency': frequency, 'now': self.context.get('now', timezone.now().date())}).data
-            for frequency, _ in Indicator.TARGET_FREQUENCIES if frequency not in [
-                Indicator.EVENT, Indicator.LOP, Indicator.MID_END
-            ]
-        }
-
-
-IPTTQSProgramSerializer = get_serializer(
-    IPTTQSMixin,
-    ProgramPeriodsMixin,
-    ProgramReportingPeriodMixin,
-    ProgramBase
-)
-
-class IPTTProgramLevelsMixin:
-    levels = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'levels'
-        ]
-
-    def get_levels(self, program):
-        levels = self._get_program_levels(program)
-        return IPTTLevelSerializer(
-            levels,
-            context={
-                'levels': levels,
-                'tiers': self._get_program_tiers(program),
-                'indicators': self._get_program_indicators(program)
-            },
-            many=True).data
-
-
-IPTTProgramLevelSerializer = get_serializer(
-    IPTTProgramLevelsMixin,
-    ProgramBase
-)
-
-class IPTTProgramFilterItemsMixin:
-    result_chain_label = serializers.SerializerMethodField()
-    tiers = serializers.SerializerMethodField()
-    sectors = serializers.SerializerMethodField()
-    indicator_types = serializers.SerializerMethodField()
-    sites = serializers.SerializerMethodField()
-    disaggregations = serializers.SerializerMethodField()
-    old_levels = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'result_chain_label',
-            'tiers',
-            'sectors',
-            'indicator_types',
-            'sites',
-            'disaggregations',
-            'old_levels',
-        ]
-
-    def get_result_chain_label(self, program):
-        """returns "Outcome Chain" or "Chaîne Résultat" for labeling the ordering filter"""
-        tier_name = self._get_result_tier(program)
-        if tier_name:
-            return _('%(tier)s Chain') % {'tier': _(tier_name)}
-        return None
-
-    def get_tiers(self, program):
-        return IPTTTierSerializer(self._get_program_tiers(program), many=True).data
-
-    def _get_program_indicator_types(self, program):
-        if hasattr(self, 'context') and 'indicator_types' in self.context:
-            return self.context['indicator_types']
-        return IndicatorType.objects.filter(indicator__program=program).values('pk', 'indicator_type')
-
-    def _get_program_sectors(self, program):
-        if hasattr(self, 'context') and 'sectors' in self.context:
-            return self.context['sectors']
-        return Sector.objects.filter(indicator__program=program).values('pk', 'sector')
-
-    def _get_program_sites(self, program):
-        if hasattr(self, 'context') and 'sites' in self.context:
-            return self.context['sites']
-        return SiteProfile.objects.filter(result__indicator__program=program).values('pk', 'name')
-
-    def _get_program_disaggregations(self, program):
-        if hasattr(self, 'context') and 'disaggregations' in self.context:
-            return self.context['disaggregations']
-        return DisaggregationType.objects.filter(indicator__program=program).values(
-            'pk', 'disaggregation_type', 'standard', 'country__country'
-        )
-
-    def _get_program_disaggregation_labels(self, program):
-        if hasattr(self, 'context') and 'disaggregation_labels' in self.context:
-            return self.context['disaggregation_labels']
-        return DisaggregationLabel.objects.filter(
-            disaggregation_type__indicator__program=program
-        ).values('pk', 'disaggregation_type_id', 'label', 'customsort')
-
-    def get_sectors(self, program):
-        return sorted({
-            v['pk']: v for v in [{'pk': sector['pk'], 'name': sector['sector']}
-                                 for sector in self._get_program_sectors(program)]
-            }.values(), key=operator.itemgetter('name'))
-
-    def get_indicator_types(self, program):
-        return sorted({
-            v['pk']: v for v in [{'pk': indicator_type['pk'], 'name': indicator_type['indicator_type']}
-                                 for indicator_type in self._get_program_indicator_types(program)]
-            }.values(), key=operator.itemgetter('name'))
-
-    def get_sites(self, program):
-        return sorted({
-            v['pk']: v for v in [{'pk': site['pk'], 'name': site['name']} for site in self._get_program_sites(program)]
-        }.values(), key=operator.itemgetter('name'))
-
-    def get_disaggregations(self, program):
-        labels = [{'pk': l['pk'], 'name': l['label'],
-                   'customsort': l['customsort'], 'disaggregation': l['disaggregation_type_id']}
-                  for l in self._get_program_disaggregation_labels(program)]
-        return sorted({
-            v['pk']: v for v in [
-                {'pk': disaggregation['pk'], 'name': disaggregation['disaggregation_type'],
-                 'country': None if disaggregation['standard'] else disaggregation['country__country'],
-                 'labels': [label for label in labels if label['disaggregation'] == disaggregation['pk']]}
-                for disaggregation in self._get_program_disaggregations(program)
-            ]
-        }.values(), key=operator.itemgetter('name'))
-
-    def get_old_levels(self, program):
-        if program.results_framework:
-            return []
-        old_level_pks = {name: pk for (pk, name) in Indicator.OLD_LEVELS}
-        return sorted([{'pk': old_level_pks[name], 'name': _(name)} for name in set(
-            i.old_level for i in self._get_program_indicators(program) if i.old_level)], key=operator.itemgetter('pk'))
-
-
-class IPTTMixin:
-    indicators = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'indicators'
-        ]
-
-    @classmethod
-    def get_for_pk(cls, program_pk):
-        program = Program.rf_aware_objects.select_related(None).prefetch_related(None).only(
-            *cls._get_query_fields()
-        ).prefetch_related(
-            models.Prefetch(
-                'indicator_set',
-                queryset=Indicator.rf_aware_objects.select_related(None).prefetch_related(None).only(
-                    'pk', 'name', 'deleted', 'program_id', 'means_of_verification', 'level_id', 'level_order',
-                    'number', 'target_frequency', 'unit_of_measure', 'unit_of_measure_type', 'baseline', 'baseline_na',
-                    'direction_of_change', 'is_cumulative', 'key_performance_indicator', 'old_level',
-                    'create_date', 'sector_id'
-                ),
-                to_attr='prefetch_indicators'
-            ),
-            models.Prefetch(
-                'level_tiers',
-                queryset=LevelTier.objects.select_related(None).only(
-                    'pk', 'name', 'program_id', 'tier_depth'
-                ),
-                to_attr='prefetch_tiers'
-            ),
-            models.Prefetch(
-                'levels',
-                queryset=Level.objects.select_related(None).only(
-                    'pk', 'name', 'parent_id', 'customsort', 'program_id'
-                ),
-                to_attr='prefetch_levels'
-            ),
-        ).get(pk=program_pk)
-        context = {
-            'indicator_types': IndicatorType.objects.select_related(None).prefetch_related(None).filter(
-                indicator__program_id=program_pk
-            ).order_by('indicator_type').values('pk', 'indicator_type', 'indicator__pk'),
-            'sites': SiteProfile.objects.select_related(None).prefetch_related(None).filter(
-                result__indicator__program_id=program_pk
-            ).order_by('name').values('pk', 'name', 'result__indicator__pk'),
-            'sectors': Sector.objects.select_related(None).prefetch_related(None).filter(
-                indicator__program_id=program_pk
-            ).order_by('sector').values('pk', 'sector', 'indicator__pk'),
-            'disaggregations': DisaggregationType.objects.select_related(None).prefetch_related(None).filter(
-                indicator__program_id=program_pk
-            ).order_by('disaggregation_type').values(
-                'pk', 'disaggregation_type', 'indicator__pk', 'standard', 'country__country'
-            ),
-            'disaggregation_labels': DisaggregationLabel.objects.select_related(None).prefetch_related(None).filter(
-                disaggregation_type__indicator__program_id=program_pk
-            ).order_by('customsort').values('pk', 'disaggregation_type_id', 'label', 'customsort').distinct(),
-            'now': timezone.now().date()
-        }
-        return cls(program, context=context)
-
-    def get_indicators(self, program):
-        indicators = self._get_program_indicators(program)
-        return IPTTIndicatorSerializer(indicators, context=self.context, many=True).data
-
-
-IPTTProgramSerializer = get_serializer(
-    IPTTMixin,
-    IPTTProgramFilterItemsMixin,
-    IPTTProgramLevelsMixin,
-    ProgramPeriodsMixin,
-    RFLevelOrderingMixin,
-    ProgramReportingPeriodMixin,
-    ProgramBase
-)
-
-class IPTTExcelMixin:
-    levels = serializers.SerializerMethodField()
-    
-    class Meta:
-        fields = [
-            'levels',
-        ]
-
-    @classmethod
-    def get_for_pk(cls, program_pk, context={}):
-        program = Program.rf_aware_objects.select_related(None).prefetch_related(None).only(
-            *cls._get_query_fields()
-        ).get(pk=program_pk)
-        return cls(program, context=context)
-   
-    def get_levels(self, program):
-        if self.context.get('level_order', False):
-            return [IPTTLevelSerializer(level) for level in self._get_levels_level_order(program)]
-        else:
-            return [IPTTLevelSerializer(level) for level in self._get_levels_chain_order(program)]
-        
-        
-        
-
-IPTTExcelProgramSerializer = get_serializer(
-    IPTTExcelMixin,
-    ProgramReportingPeriodMixin,
-    ProgramBase
-)
-
-class IPTTReportSerializer(serializers.Serializer):
-    """Serializer for an entire IPTT Report - contains report-level data and methods to instance sub-serializers"""
-    report_title = serializers.SerializerMethodField()
-    program_name = serializers.SerializerMethodField()
-    report_date_range = serializers.SerializerMethodField()
-    lop_period = serializers.SerializerMethodField()
-    periods = serializers.SerializerMethodField()
-    level_rows = serializers.SerializerMethodField()
-    blank_level_rows = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'report_title',
-            'program_name',
-            'report_date_range',
-            'lop_period',
-            'periods',
-            'level_rows',
-            'blank_level_rows'
-        ]
-    
-    @property
-    def filename(self):
-        return f"{self.report_name} {l10n_date_medium(timezone.localtime().date(), decode=True)}.xlsx"
-
-    @classmethod
-    def load_program_data(cls, program_pk, program_context={}):
-        return IPTTExcelProgramSerializer.get_for_pk(program_pk, context=program_context)
-
-    @classmethod
-    def load_indicator_data(cls, indicator_context={}, indicator_filters={}):        
-        indicators = Indicator.rf_aware_objects.select_related('program').prefetch_related(None).only(
-            'pk', 'name', 'deleted', 'program_id', 'means_of_verification', 'level_id', 'level_order',
-            'number', 'target_frequency', 'unit_of_measure', 'unit_of_measure_type', 'baseline', 'baseline_na',
-            'direction_of_change', 'is_cumulative', 'key_performance_indicator', 'old_level',
-            'create_date', 'sector_id'
-        ).filter(**indicator_filters).order_by().distinct()
-        return IPTTExcelIndicatorSerializer(indicators, context=indicator_context, many=True)
-
-    @classmethod
-    def get_indicator_filters(cls, filters, program_pk=None):
-        indicator_filters = {}
-        if program_pk:
-            indicator_filters['program'] = program_pk
-        if filters.get('sectors', None):
-            indicator_filters['sector__in'] = filters.get('sectors')
-        if filters.get('types', None):
-            indicator_filters['indicator_type__in'] = filters.get('types')
-        if filters.get('indicators', None):
-            indicator_filters['pk__in'] = filters.get('indicators')
-        if filters.get('disaggregations', None):
-            indicator_filters['disaggregation__in'] = filters.get('disaggregations')
-        if filters.get('sites', None):
-            indicator_filters['result__site__in'] = filters.get('sites')
-        return indicator_filters
-
-    @classmethod
-    def get_context(cls, program_pk, frequencies, filters={}):
-        if type(frequencies) == int:
-            frequencies = [frequencies]
-        context = {}
-        program_context = {}
-        if filters.get('groupby', None) == 2:
-            program_context['level_order'] = True
-        context['program'] = cls.load_program_data(program_pk, program_context=program_context).data
-        context['frequencies'] = [int(frequency) for frequency in frequencies]
-        indicator_context = {
-            'indicator_types': IndicatorType.objects.select_related(None).prefetch_related(None).filter(
-                indicator__program_id=program_pk
-            ).order_by('indicator_type').values('pk', 'indicator_type', 'indicator__pk'),
-            'sites': SiteProfile.objects.select_related(None).prefetch_related(None).filter(
-                result__indicator__program_id=program_pk
-            ).order_by('name').values('pk', 'name', 'result__indicator__pk'),
-            'sectors': Sector.objects.select_related(None).prefetch_related(None).filter(
-                indicator__program_id=program_pk
-            ).order_by('sector').values('pk', 'sector', 'indicator__pk'),
-            'disaggregations': DisaggregationType.objects.select_related(None).prefetch_related(None).filter(
-                indicator__program_id=program_pk
-            ).order_by('disaggregation_type').values(
-                'pk', 'disaggregation_type', 'indicator__pk', 'standard', 'country__country'
-            ),
-            'disaggregation_labels': DisaggregationLabel.objects.select_related(None).prefetch_related(None).filter(
-                disaggregation_type__indicator__program_id=program_pk
-            ).order_by('customsort').values('pk', 'disaggregation_type_id', 'label', 'customsort').distinct(),
-            'now': timezone.now().date(),
-        }
-        indicator_filters = cls.get_indicator_filters(filters, program_pk)
-        context['indicators'] = cls.load_indicator_data(
-            indicator_context=indicator_context,
-            indicator_filters=indicator_filters
-        ).data
-        return context
-
-    def get_report_title(self, obj):
-        return _('Indicator Performance Tracking Report')
-
-    def get_program_name(self, obj):
-        return self.context['program']['name']
-
-    @property
-    def reporting_periods(self):
-        return (
-            dateutil.parser.isoparse(self.context['program']['reporting_period_start_iso']).date(),
-            dateutil.parser.isoparse(self.context['program']['reporting_period_end_iso']).date()
-        )
-
-    def get_report_date_range(self, obj):
-        return u'{} – {}'.format(
-            l10n_date_medium(self.reporting_periods[0], decode=True),
-            l10n_date_medium(self.reporting_periods[1], decode=True)
-        )
-
-    def get_lop_period(self, obj):
-        return Period.lop_period()
-
-    def get_periods(self, obj):
-        frequencies = self.context['frequencies']
-        periods = {}
-        if Indicator.LOP in frequencies:
-            frequencies = [f for f in frequencies if f is not Indicator.LOP]
-            periods[Indicator.LOP] = []
-        return {
-            **periods,
-            **{frequency: [
-                Period(frequency, period_dict, tva=self.is_tva)
-                for period_dict in PeriodicTarget.generate_for_frequency(frequency)(
-                    *self.reporting_periods
-                )] for frequency in  frequencies
-            }
-        }
-
-    def _get_frequency_indicators(self, frequency, level_pk=None):
-        return [
-            indicator for indicator in self.context['indicators']
-            if indicator['level_pk'] == level_pk and indicator['target_frequency'] == frequency
-            ]
-
-    def _get_level_rows_for_frequency(self, frequency):
-        for level in self.context['program']['levels']:
-            yield {
-                'level': level.data,
-                'indicators': self._get_frequency_indicators(frequency, level_pk=level.data['pk'])
-            }
-
-    def _get_blank_level_row_for_frequency(self, frequency):
-        return {
-            'level': {'display_name': _('Indicators unassigned to a results framework level')},
-            'indicators': self._get_frequency_indicators(frequency, level_pk=None)
-        }
-
-    def get_level_rows(self, obj):
-        return {
-            frequency: self._get_level_rows_for_frequency(frequency)
-            for frequency in self.context['frequencies']
-        }
-
-    def get_blank_level_rows(self, obj):
-        return {
-            frequency: self._get_blank_level_row_for_frequency(frequency)
-            for frequency in self.context['frequencies']
-        }
-
-class IPTTTPReportSerializer(IPTTReportSerializer):
-    is_tva = False
-    @property
-    def report_name(self):
-        return _("IPTT Actuals only report")
-
-    @classmethod
-    def load_report(cls, program_pk, frequency=Indicator.MONTHLY, filters={}):
-        context = cls.get_context(program_pk, [frequency,], filters=filters)
-        return cls({}, context=context)
-
-    def _get_frequency_indicators(self, frequency, level_pk=None):
-        return [indicator for indicator in self.context['indicators'] if indicator['level_pk'] == level_pk]
-
-class IPTTTVAReportSerializer(IPTTReportSerializer):
-    is_tva = True
-    @property
-    def report_name(self):
-        return _("IPTT TvA report")
-
-    @classmethod
-    def load_report(cls, program_pk, frequency=Indicator.LOP, filters={}):
-        context = cls.get_context(program_pk, [frequency,], filters=filters)
-        return cls({}, context=context)
-
-class IPTTFullReportSerializer(IPTTReportSerializer):
-    is_tva = True
-    frequencies = [
-        frequency for frequency, name in Indicator.TARGET_FREQUENCIES
-        if frequency is not Indicator.EVENT
-    ]
-
-    @property
-    def report_name(self):
-        return _("IPTT TvA full program report")
-
-    @classmethod
-    def load_report(cls, program_pk, filters={}):
-        context = cls.get_context(program_pk, cls.frequencies, filters=filters)
-        return cls({}, context=context)
